Broke tracer.rs into multiple files
